+++ src/main/java/controllers/ListarBienPorCodigo.java
@@ -11,7 +11,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class ListarBienPorCodigo {
-    public static Bien obtenerBienPorCodigo(int codigo) {
+    public static Bien obtenerBienPorCodigo(long codigo) {
         Connection conn = null;
         Bien bien = null;
         try {
@@ -22,11 +22,11 @@ public class ListarBienPorCodigo {
             "INNER JOIN MA_Dependencias d ON b.FK_Dependencia = d.PK_idDependencia " +
             "WHERE b.PK_Codigo = ?";
             PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, codigo);
+            ps.setLong(1, codigo);
             ResultSet rs = ps.executeQuery();
             if (rs.next()) {
                 bien = new Bien();
-                bien.setCodigo(rs.getInt("PK_Codigo"));
+                bien.setCodigo(rs.getLong("PK_Codigo"));
                 bien.setPlaca(rs.getInt("placa"));
                 bien.setNombre(rs.getString("nombre"));
                 bien.setDescripcion(rs.getString("descripcion"));
@@ -36,7 +36,7 @@ public class ListarBienPorCodigo {
                 Usuario usuario = new Usuario();
                 usuario.setUsuario(rs.getString("usuario"));
                 bien.setUsuario(usuario);    
-                bien.setValor(rs.getInt("valor"));
+                bien.setValor(rs.getLong("valor"));
                 bien.setEstado(rs.getString("estado"));
             }
             rs.close();
